refactor: organize to module
